pyproject.toml
[tool.poetry]
name = "boe-risk-monitoring"
version = "0.0.1"
description = "Package for identifying risks of bank firms in line with Bank of England criteria"
authors = [
	"Zahin Rahim <zahinrahim03@gmail.com>",
]
readme = "README.md"
packages = [
	{include = "boe_risk_monitoring"}
]

[tool.poetry.dependencies]
python=">=3.11"
aiohttp=">=3.12"
google-genai=">=1.19"
openai=">=1.84"
pydantic=">=2.11"
pymupdf=">=1.26"
python-dotenv=">=1.1"
pandas = "^2.3.0"
langchain-core = "^0.3.64"
tiktoken = "^0.9.0"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
